+++ core/context_processors.py
@@ -1,7 +1,8 @@
 from django.conf import settings
+from core.models import ThemeConfig
 
 def global_settings(request):
     return {
         'RUBRIC_MODE': settings.RUBRIC_MODE,
-        # add as many as you like
+        'active_theme': ThemeConfig.get_active(),
     }